Extract and test getIcons function
