+++ frontend/src/metabase/collections/components/CollectionSidebar/Bookmarks/getIcon.ts
@@ -1,0 +1,15 @@
+import { BookmarkableEntities } from "metabase-types/api";
+
+export function getIcon(display: string, type: BookmarkableEntities) {
+  if (display) {
+    return display;
+  }
+
+  const icons = {
+    card: "grid",
+    collection: "folder",
+    dashboard: "dashboard",
+  };
+
+  return icons[type];
+}
